move result class to a new file
- Add comments to main
