+++ bin/main.rb
@@ -1,9 +1,12 @@
 require_relative '../lib/scraper'
+require_relative '../lib/result'
+
 hash = { 0 => 'employer\'s name', 1 => 'job title', 2 => 'city\'s name', 3 => 'year between 2012 and 2020' }
 instructions = 'The value can be alphanumeric and have one space between words.'
 all_results = []
 rewrite = 'Please write a valid option'
 
+# Verify the input for the four main options
 def valid_option(number, rewrite)
   until (0..4).include? number
     puts rewrite
@@ -12,6 +15,7 @@ def valid_option(number, rewrite)
   number
 end
 
+# Calls methods depending on the value of valid_result
 def name_year(scraper, valid_result, hash, instructions)
   if valid_result < 3
     valid_name(valid_result, scraper, hash, instructions)
@@ -26,6 +30,7 @@ def name_year(scraper, valid_result, hash, instructions)
   end
 end
 
+# Verify the input to save it in the instance variable form
 def valid_name(number, obj, hash, string)
   input = ''
   loop do
@@ -37,6 +42,7 @@ def valid_name(number, obj, hash, string)
   obj.form[number] = input =~ /\w/ ? input : ''
 end
 
+# Verify the input to save it in the instance variable form as the year
 def valid_year(obj, hash)
   input = ''
   loop do
@@ -47,11 +53,13 @@ def valid_year(obj, hash)
   obj.form[3] = input =~ /\w/ ? input : obj.form[3]
 end
 
+# Displays all the input provided for the user
 def display_input(scraper)
   puts "\nGood! These are the values \n\n Employer: #{scraper.form[0]} \n Job's Title: #{scraper.form[1]}
  City: #{scraper.form[2]} \n Year: #{scraper.form[3]}"
 end
 
+# Displays the values of each Result instance
 def display_values(all_results)
   all_results.each do |x|
     puts " Employer: #{x.employer} \n Job Title: #{x.job_title} \n Base Salary: #{x.base_salary}
@@ -60,6 +68,7 @@ def display_values(all_results)
   end
 end
 
+# The user selects to save the search results or to discard them
 def display_discard(findings, url, all_results, rewrite)
   loop do
     answer = gets.chomp
@@ -75,6 +84,7 @@ def display_discard(findings, url, all_results, rewrite)
   end
 end
 
+# Add values to Result instances and add then to all_results
 def save_results(findings, url, all_results)
   findings.times do |idx|
     obj = Result.new
@@ -83,6 +93,7 @@ def save_results(findings, url, all_results)
   end
 end
 
+# This method closes the program or let restart the main loop
 def close_scraper(rewrite)
   puts 'Would you like to do another search? (y) to continue (n) to exit.'
   loop do
@@ -110,6 +121,7 @@ puts "\nTo do the search it\'s necessary to apply al least one of these filters:
 By default, the scrapper does the search including all the available years, which are
 from 2012 to 2020."
 
+# Main loop
 loop do
   puts "\nPlease write the number of the option.
   In case you want to add more that one filter,
